requirements.txt
Flask==2.2.5
Flask-Mail==0.9.1
twilio==8.0.0
mysql-connector-python==8.0.32
python-dotenv==1.0.0
Werkzeug==2.2.3
pycryptodome
